[4.0.0] Export everything by default
