+++ gen/generator.toml
@@ -1,7 +1,7 @@
 [general]
 library_name = "libfreetype"
-output_file_path = "../src/LibFreeType.jl"
-module_name = "LibFreeType"
+output_file_path = "../src/FreeType.jl"
+module_name = "FreeType"
 jll_pkg_name = "FreeType2_jll"
 export_symbol_prefixes = ["FREETYPE_", "FT_", "ft_", "TT_", "TTAG_", "CFF_", "T1_", "CID_", "PS_", "t1_"]
 prologue_file_path = "./prologue.jl"
